unifying all single gpu rf test with w/o fil
